socket: use accept4 to preserve flags in child socket
when the tracee calls the `accept4` syscall to accept an incoming TCP
connection, it will pass a `flags` argument that specifies whether the socket
should be non-blocking. we handle this correctly in outgoing connections and
listening sockets. but for some reason, in the intercept handler for `accept4`,
I used the older `accept` syscall when creating the child's new socket
(`accept` does not support non-blocking syscalls; `accept(fd)` is the same as
`accept4(fd, 0)`).

so a tracee may call `read(fd)` on that socket even if there's nothing left to
read because it expects such calls to fail with `EAGAIN`. but since we never
marked the socket as non-blocking, the call would block forever.

this bug was particularly insidious because the websocket hanging has nothing
to do with the accepted socket. when a HTTP requests gets upgraded to a
websocket with a status 101, the websocket will still use the same TCP
connection. we've already accepted the TCP connection and the tracee has
already responded with 101, so our `accept4` handler surely can't be why the
_websocket_ was hanging, right? nope. because of the way the Go handles I/O
internally for perf reasons, `(net.Conn).Read()` calls don't actually translate
to the `read` system call. instead, Go relies on an internal event poll to know
when something is readable and calls `read` only when necessary; if the event
poll hasn't fired yet, there will be nothing to read, so why waste a syscall
and suffer from the context switch?

that still doesn't explain why the websocket would hang tho because there _are_
websocket messages to read from the wire! I mean, I saw the bytes on wireshark!
what actually caused the failure was one more level away: when the Go HTTP
server successfully parses a single request, it starts a background goroutine
to read the next request even if the current request's response hasn't been
written because of an obscure HTTP/1.x feature called pipelining. almost no
HTTP client uses it except for `apt-get` (sometimes).

so when the first `/tunnel` request finishes, the Go HTTP server implementation
starts a goroutine to fetch the next request's bytes. that's normally fine
because there's usually another request there or the client goes away. but we
upgrade the connection to a websocket. that's also normally fine because Go
uses `SetReadDeadline` to cancel all pending `Read` calls that might be
inflight. you can't really cancel syscalls without closing the socket, so how
do they do that? with non-blocking sockets and event poll notifications, of
course. so we tell the background read to stop and it calls `abortPendingRead`,
but the internal poll handler became a zombie goroutine ages ago because it
called `read` on a completely unrelated socket expecting it to return
immediately, so there's no one around to actually do the complete the
cancellation.

so the websocket connection never becomes active sometimes even though it looks
fine from the other side of the connection. so `subtrace run -- curl` works
fine and tries to pushes trace events at exit but `subtrace run -- control`
zombifies the websocket tunnel and the worker never sees the data.